requirements.txt
mysql-connector-python==8.0.x
python-dotenv==1.0.x
